[s390] pci: don't cleanup in arch_setup_msi_irqs
Message-id: <1507208888-28130-2-git-send-email-brueckner@redhat.com>
Patchwork-id: 189297
O-Subject: [RHEL7.5 PATCH 01/10] [s390] s390/pci: don't cleanup in arch_setup_msi_irqs
Bugzilla: 1486403
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Cornelia Huck <cohuck@redhat.com>

Description
-----------
commit 795818e8bf17dbc791764ba9fb723278ee934676
Author: Sebastian Ott <sebott@linux.vnet.ibm.com>
Commit: Martin Schwidefsky <schwidefsky@de.ibm.com>

    s390/pci: don't cleanup in arch_setup_msi_irqs

    After failures in arch_setup_msi_irqs common code calls
    arch_teardown_msi_irqs. Thus, remove cleanup code from
    arch_setup_msi_irqs.


Notes
-----
This backport include some changes due to missing generic hardirqs
introduced with commit 1f44a225777e40fd9a945b09f958052c47494e1e
"s390: convert interrupt handling to use generic hardirq".

Bugzilla
--------
BZ 1486403
https://bugzilla.redhat.com/show_bug.cgi?id=1486403

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version v4.13-rc1
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux.git;a=commitdiff;h=795818e8bf17dbc791764ba9fb723278ee934676

Brew
----
https://brewweb.devel.redhat.com/taskinfo?taskID=14174257

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

+++ arch/s390/include/asm/pci.h
@@ -115,7 +115,7 @@ struct zpci_dev {
 	struct zdev_irq_map *irq_map;
 	struct msi_map *msi_map;
 	struct airq_iv *aibv;		/* adapter interrupt bit vector */
-	unsigned int	aisb;		/* number of the summary bit */
+	unsigned long	aisb;		/* number of the summary bit */
 
 	/* DMA stuff */
 	unsigned long	*dma_table;
